requirements.txt
# Deep Learning Framework
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# PDF Processing with Marker
marker-pdf[full]>=0.2.0

# PDF Processing (fallback)
PyMuPDF>=1.23.0
pillow>=9.0.0

# NLP and transformers
transformers>=4.35.0
sentence-transformers>=2.2.2
spacy>=3.7.0

# Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0

# Web Framework
flask>=2.3.0
flask-restx>=1.1.0
flask-cors>=4.0.0

# Data Validation
pydantic>=2.0.0

# Utilities
numpy>=1.24.0,<2.0
scipy>=1.10.0
scikit-learn>=1.3.0
requests>=2.28.0
python-multipart>=0.0.6

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Core ML Libraries
huggingface-hub>=0.20.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Additional utilities
psutil>=5.9.0